Refactor all nodes into lib/ folder
