+++ datamodel/nodes/lib/functional_operators.py
