+++ datamodel/nodes/lib/strings.py
